pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-team"
version = "0.1.0"
description = "CLI for building and managing orchestrated teams of Hermes agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["hermes", "agents", "orchestration", "ai", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
orchestrator = "orchestrator.cli:cli"

[project.urls]
Homepage = "https://github.com/Nitsua365/hermes-team"
Repository = "https://github.com/Nitsua365/hermes-team"
Issues = "https://github.com/Nitsua365/hermes-team/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator*"]

[tool.setuptools.package-data]
"orchestrator.scaffold" = [
    "Dockerfile",
    "entrypoint.sh",
    "hermes-team.yaml",
    "docker-compose.yml",
    "tools/*",
    "skills/**/*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["orchestrator"]
omit = ["tests/*"]
